[WIP] audio(8): Initial revision
audio(8) is a new program that lists and modifies audio device
properties, in similar fashion to mixer(8).

It works using a control/variable-driven interface; these controls are
defined in `dev_ctls` and `chan_ctls`. Most of them are read-only, but
some of them can also be modified, see which ones have a non-NULL `mod`
field.

Sponsored by:	The FreeBSD Foundation
MFC after:	1 week