sound: Fix hot-unload page fault
This panic is easily reproduced by spawning many channels, and unloading
the module right after. The panic messages usually state that we take a
page fault in destroy_devl() with the cdev lock held. My theory is that
we go to sleep somewhere in destroy_devl(), and by the time we wake up,
pcm_killchans() will have freed resources that destroy_devl() will try
to access, resulting in a page fault.

Sponsored by:	The FreeBSD Foundation
MFC after:	2 days